refactor(core): extract resolution planner
